+++ src/main/java/com/example/springexample/services/lib/CRUDService.java
@@ -3,9 +3,9 @@ package com.example.springexample.services.lib;
 import java.util.List;
 
 public interface CRUDService<T> {
-    public T getById(Integer id);
-    public List<T> getTop(Integer item, Integer count);
+    public T getById(Long id);
+    public List<T> getTop(Long item, Integer count);
     public void create (T item);
     public void update(T item);
-    public void delete(Integer id);
+    public void delete(Long id);
 }
